+++ client/src/app/(core)/project/create/project_creation_form.tsx
@@ -1,0 +1,56 @@
+import { useForm, SubmitHandler } from 'react-hook-form';
+
+import './project_form.css';
+
+interface IFormInput {
+	firstName: string;
+	lastName: string;
+	age: number;
+	example: string;
+}
+
+export default function ProjectForm() {
+	const {
+		register,
+		handleSubmit,
+		watch,
+		formState: { errors },
+	} = useForm<IFormInput>();
+
+	const onSubmit = (data: IFormInput) => {
+		alert(JSON.stringify(data));
+	}; // your form submit function which will invoke after successful validation
+
+	console.log(watch('example')); // you can watch individual input by pass the name of the input
+
+	return (
+		<form onSubmit={handleSubmit(onSubmit)}>
+			<label>First Name</label>
+			<input
+				{...register('firstName', {
+					required: true,
+					maxLength: 20,
+					pattern: /^[A-Za-z]+$/i,
+				})}
+			/>
+			{errors?.firstName?.type === 'required' && <p>This field is required</p>}
+			{errors?.firstName?.type === 'maxLength' && (
+				<p>First name cannot exceed 20 characters</p>
+			)}
+			{errors?.firstName?.type === 'pattern' && (
+				<p>Alphabetical characters only</p>
+			)}
+			<label>Last Name</label>
+			<input {...register('lastName', { pattern: /^[A-Za-z]+$/i })} />
+			{errors?.lastName?.type === 'pattern' && (
+				<p>Alphabetical characters only</p>
+			)}
+			<label>Age</label>
+			<input {...register('age', { min: 18, max: 99 })} />
+			{errors.age && (
+				<p>You Must be older then 18 and younger then 99 years old</p>
+			)}
+			<input type="submit" />
+		</form>
+	);
+}
